Refactoring and documentation
Just did some abstraction and documentation.
